refactor: edit widget field dir structure
