+++ apps/web/src/common/modules/widgets/_widget-fields/category-by/type.ts
@@ -1,0 +1,12 @@
+export interface CategoryByValue {
+    value: string;
+    count: number;
+}
+
+export interface CategoryByOptions {
+    dataTarget?: string;
+    max?: number;
+    defaultMaxCount: number;
+    defaultIndex?: number;
+    excludeDateField?: boolean;
+}
