+++ apps/web/src/common/modules/widgets/_widget-fields/format-rules/type.ts
@@ -1,0 +1,16 @@
+import type { FORMAT_RULE_TYPE } from '@/common/modules/widgets/_constants/widget-field-constant';
+
+export interface FormatRulesValue {
+    text?: string;
+    threshold?: number;
+    color: string;
+}
+
+export type FormatRulesType = typeof FORMAT_RULE_TYPE[keyof typeof FORMAT_RULE_TYPE];
+
+export interface FormatRulesOptions {
+    formatRulesType: FormatRulesType;
+    description?: string;
+    default?: FormatRulesValue[];
+    baseColor?: string;
+}
