+++ apps/web/src/common/modules/widgets/_widgets/gauge/Gauge.vue
@@ -27,12 +27,13 @@ import {
     getWidgetDateRange,
 } from '@/common/modules/widgets/_helpers/widget-date-helper';
 import { getFormattedNumber } from '@/common/modules/widgets/_helpers/widget-helper';
+import type { FormatRulesValue } from '@/common/modules/widgets/_widget-fields/format-rules/type';
 import type { DateRange } from '@/common/modules/widgets/types/widget-data-type';
 import type {
     WidgetProps, WidgetEmit,
     WidgetExpose,
 } from '@/common/modules/widgets/types/widget-display-type';
-import type { FormatRulesValue, NumberFormatValue } from '@/common/modules/widgets/types/widget-field-value-type';
+import type { NumberFormatValue } from '@/common/modules/widgets/types/widget-field-value-type';
 
 import { gray } from '@/styles/colors';
